+++ lib/view_model/product_view_model.dart
@@ -5,9 +5,10 @@ import 'package:riverpod/riverpod.dart';
 
 class ProductViewModel extends StateNotifier<List<List<dynamic>>> {
   ProductViewModel() : super([]) {
-    _loadCSV();  // 생성자에서 초기화 작업 수행
+    _loadCSV();
   }
 
+  //csv 파일을 읽어와서 헤더 제외하고 리스트로 변환
   Future<void> _loadCSV() async {
     try {
       NumberFormat numberFormat = NumberFormat('#,###');
@@ -25,8 +26,9 @@ class ProductViewModel extends StateNotifier<List<List<dynamic>>> {
           row[5] = '${numberFormat.format(row[5]).toString()}원';
         }
       }
-
-      state = listData.sublist(2);  // 첫 두 개의 헤더를 제외하고 상태 업데이트
+    
+      // 첫 두 개의 헤더를 제외하고 상태 업데이트
+      state = listData.sublist(2);
     } catch (e) {
       print("Error loading CSV: $e");
     }
@@ -34,26 +36,30 @@ class ProductViewModel extends StateNotifier<List<List<dynamic>>> {
 
   // 상품 삭제 메서드
   Future<void> deleteProduct(int index) async {
-    List<List<dynamic>> updatedState = List.from(state);  
-    updatedState.removeAt(index); 
-    state = updatedState;  
+    List<List<dynamic>> updatedState = List.from(state);
+    updatedState.removeAt(index);
+    state = updatedState;
   }
+
   //상품 좋아요 기능
-  Future<void> updateLike(int index)async{
-    List<List<dynamic>> updatedState = List.from(state);  
-    updatedState[index][7] ++;
+  Future<void> updateLike(int index) async {
+    List<List<dynamic>> updatedState = List.from(state);
+    updatedState[index][7]++;
     updatedState[index].add(true);
-    state = updatedState;  
+    state = updatedState;
   }
-  Future<void> updateUnlike(int index)async{
-    List<List<dynamic>> updatedState = List.from(state);  
-    updatedState[index][7] --;
+
+  //상품 좋아요 취소 기능
+  Future<void> updateUnlike(int index) async {
+    List<List<dynamic>> updatedState = List.from(state);
+    updatedState[index][7]--;
     updatedState[index].remove(true);
-    state = updatedState;  
+    state = updatedState;
   }
 }
 
 // ProductViewModel을 제공하는 Provider
-final productViewModelProvider = StateNotifierProvider<ProductViewModel, List<List<dynamic>>>((ref) {
+final productViewModelProvider =
+    StateNotifierProvider<ProductViewModel, List<List<dynamic>>>((ref) {
   return ProductViewModel();
 });
